+++ crates/services/producer/src/block_producer/tests.rs
@@ -2,32 +2,62 @@
 
 use crate::{
     block_producer::{
-        gas_price::{GasPriceProvider, MockConsensusParametersProvider},
-        Bytes32, Error,
+        gas_price::{
+            GasPriceProvider,
+            MockConsensusParametersProvider,
+        },
+        Bytes32,
+        Error,
     },
     mocks::{
-        FailingMockExecutor, MockDb, MockExecutor, MockExecutorWithCapture, MockRelayer,
+        FailingMockExecutor,
+        MockDb,
+        MockExecutor,
+        MockExecutorWithCapture,
+        MockRelayer,
         MockTxPool,
     },
-    Config, Producer,
+    Config,
+    Producer,
 };
 use fuel_core_producer as _;
 use fuel_core_types::{
     blockchain::{
-        block::{Block, CompressedBlock, PartialFuelBlock},
-        header::{ApplicationHeader, ConsensusHeader, PartialBlockHeader},
+        block::{
+            Block,
+            CompressedBlock,
+            PartialFuelBlock,
+        },
+        header::{
+            ApplicationHeader,
+            ConsensusHeader,
+            PartialBlockHeader,
+        },
         primitives::DaBlockHeight,
     },
     fuel_tx,
-    fuel_tx::{field::InputContract, ConsensusParameters, Mint, Script, Transaction},
+    fuel_tx::{
+        field::InputContract,
+        ConsensusParameters,
+        Mint,
+        Script,
+        Transaction,
+    },
     fuel_types::BlockHeight,
     services::executor::Error as ExecutorError,
     tai64::Tai64,
 };
-use rand::{rngs::StdRng, Rng, SeedableRng};
+use rand::{
+    rngs::StdRng,
+    Rng,
+    SeedableRng,
+};
 use std::{
     collections::HashMap,
-    sync::{Arc, Mutex},
+    sync::{
+        Arc,
+        Mutex,
+    },
 };
 
 pub struct MockProducerGasPrice {
@@ -560,7 +590,10 @@ mod produce_and_execute_block_txpool {
 }
 
 use fuel_core_types::fuel_tx::field::MintGasPrice;
-use proptest::{prop_compose, proptest};
+use proptest::{
+    prop_compose,
+    proptest,
+};
 
 prop_compose! {
     fn arb_block()(height in 1..255u8, da_height in 1..255u64, gas_price: u64, coinbase_recipient: [u8; 32], num_txs in 0..100u32) -> Block {